+++ src/js/vre/chapter31/components/IntroductionPage.jsx
@@ -19,7 +19,7 @@ class IntroductionPage extends React.Component {
         <SaveInProgressIntro
           buttonOnly
           pageList={this.props.route.pageList}
-          startText="Start the Chapter 36 application"
+          startText="Start the Chapter 31 application"
           {...this.props.saveInProgressActions}
           {...this.props.saveInProgress}/>
         <div className="omb-info--container" style={{ paddingLeft: '0px' }}>
