--- src/js/vre/chapter36/labels.js
@@ -1,8 +1,0 @@
-export const dischargeTypeLabels = {
-  honorable: 'Honorable',
-  general: 'General',
-  other: 'Other Than Honorable',
-  'bad-conduct': 'Bad Conduct',
-  dishonorable: 'Dishonorable',
-  undesirable: 'Undesirable'
-};
